backend/requirements.txt
fastapi
uvicorn
motor
pydantic
pydantic-settings
python-dotenv
passlib
bcrypt
python-jose
python-multipart
pymongo
requests
pandas
google-generativeai
email-validator
pytest
httpx
PyJWT
feedparser
newspaper3k
lxml_html_clean